Refactor the actual layout code into Font
It can be useful to lay things out without drawing them, e.g. querying
the total size of a region of rendered text.
